pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Sayam Gupta <gupta.sayam2006@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
mistralai = "^1.1.0"
google-generativeai = "^0.8.3"
pillow = "^11.0.0"
pyrebase4 = "^4.8.0"
firebase-admin = "^6.5.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
